requirements.txt
langchain
langgraph
langchain-openai
langchain-community
faiss-cpu
python-dotenv
